Add test code

+++ junit/4.8/project/src/main/java/org/test/Calc.java
@@ -7,8 +7,8 @@ public class Calc {
 		return a+b;
 	}
 	
-	public int divide(int a, int b){
+	public float divide(int a, int b){
 		
-		return a / b;
+		return (float)a / (float)b;
 	}
 }
